continue processing work after reaping
If there is more work in the todo queue than threads
available in the pool to process the work, but somehow
the threads are no longer alive, the reaper appropriately
comes along and reaps them.  However, if there are remaining
work items to do, they are not scheduled to be run, but
sit there not being processed.  They are only scheduled to be
processed when a new item is pushed into the queue.