resolve waiting counter problem
if a thread is in the wait state, but transitions into
a killed state the waiting counter will not be decremented
leaving the pool to believe there is a thread available
and waiting to complete work when that is not the case.
Introduce a begin..ensure block so that any thread exception
raised while in the wait state properly balances the
waiting counter.