add missing init? and shutdown methods
Puma adjusted their SSL support for MRI Ruby, but it
appears as if they did not keep the JRuby version up
to date.  Here is the offending commit:
https://github.com/puma/puma/commit/46416cb49ed2f16614f019cee969bb8f5d0a6146
note that it adds 2 methods.  'init?' and 'shutdown'.
We add the same methods for the JRuby SSL implementation
to keep parity with the MRI Ruby implementation.